backend/IpApp/requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.6
pymongo==3.11.0
python-dotenv==0.12.0
requests==2.25.0
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
